refactor(memory): remove FTS, LIKE search, query expansion, and LLM reranker
