+++ assistant/src/__tests__/memory-lifecycle-e2e.test.ts
@@ -203,7 +203,7 @@ describe("Memory lifecycle E2E regression", () => {
     db.run("DELETE FROM memory_embeddings");
     db.run("DELETE FROM memory_summaries");
     db.run("DELETE FROM memory_items");
-    db.run("DELETE FROM memory_segment_fts");
+
     db.run("DELETE FROM memory_segments");
     db.run("DELETE FROM messages");
     db.run("DELETE FROM conversations");
