+++ assistant/src/__tests__/memory-regressions.experimental.test.ts
@@ -88,7 +88,7 @@ describe("Memory regressions (experimental)", () => {
     db.run("DELETE FROM memory_embeddings");
     db.run("DELETE FROM memory_summaries");
     db.run("DELETE FROM memory_items");
-    db.run("DELETE FROM memory_segment_fts");
+
     db.run("DELETE FROM memory_segments");
     db.run("DELETE FROM messages");
     db.run("DELETE FROM conversations");
