+++ assistant/src/__tests__/memory-upsert-concurrency.test.ts
@@ -110,7 +110,6 @@ function resetTables() {
   db.run("DELETE FROM memory_embeddings");
   db.run("DELETE FROM memory_summaries");
   db.run("DELETE FROM memory_items");
-  db.run("DELETE FROM memory_segment_fts");
   db.run("DELETE FROM memory_segments");
   db.run("DELETE FROM memory_jobs");
   db.run("DELETE FROM messages");
