+++ assistant/src/memory/conversation-crud.ts
@@ -503,19 +503,6 @@ export function clearAll(): { conversations: number; messages: number } {
   // triggers so that the subsequent base-table DELETEs don't also fail
   // (SQLite triggers are atomic with the triggering statement, so a
   // corrupted FTS table would roll back every base-table DELETE).
-  let segmentFtsCorrupted = false;
-  try {
-    rawExec("DELETE FROM memory_segment_fts");
-  } catch (err) {
-    log.warn(
-      { err },
-      "clearAll: failed to clear memory_segment_fts — dropping triggers so base-table cleanup can proceed",
-    );
-    rawExec("DROP TRIGGER IF EXISTS memory_segments_ai");
-    rawExec("DROP TRIGGER IF EXISTS memory_segments_ad");
-    rawExec("DROP TRIGGER IF EXISTS memory_segments_au");
-    segmentFtsCorrupted = true;
-  }
   rawExec("DELETE FROM memory_item_sources");
   rawExec("DELETE FROM memory_segments");
   rawExec("DELETE FROM memory_items");
@@ -548,21 +535,6 @@ export function clearAll(): { conversations: number; messages: number } {
   // DELETEs have completed. Dropping the virtual table clears the corruption,
   // and recreating it + triggers means subsequent writes maintain FTS
   // consistency without requiring a daemon restart.
-  if (segmentFtsCorrupted) {
-    rawExec("DROP TABLE IF EXISTS memory_segment_fts");
-    rawExec(
-      `CREATE VIRTUAL TABLE IF NOT EXISTS memory_segment_fts USING fts5(segment_id UNINDEXED, text)`,
-    );
-    rawExec(
-      `CREATE TRIGGER IF NOT EXISTS memory_segments_ai AFTER INSERT ON memory_segments BEGIN INSERT INTO memory_segment_fts(segment_id, text) VALUES (new.id, new.text); END`,
-    );
-    rawExec(
-      `CREATE TRIGGER IF NOT EXISTS memory_segments_ad AFTER DELETE ON memory_segments BEGIN DELETE FROM memory_segment_fts WHERE segment_id = old.id; END`,
-    );
-    rawExec(
-      `CREATE TRIGGER IF NOT EXISTS memory_segments_au AFTER UPDATE ON memory_segments BEGIN DELETE FROM memory_segment_fts WHERE segment_id = old.id; INSERT INTO memory_segment_fts(segment_id, text) VALUES (new.id, new.text); END`,
-    );
-  }
   if (messagesFtsCorrupted) {
     rawExec("DROP TABLE IF EXISTS messages_fts");
     rawExec(
