+++ assistant/src/memory/job-handlers/index-maintenance.ts
@@ -2,7 +2,7 @@ import { eq, like } from "drizzle-orm";
 
 import { getConfig } from "../../config/loader.js";
 import { getLogger } from "../../util/logger.js";
-import { getDb, rawExec } from "../db.js";
+import { getDb } from "../db.js";
 import { selectedBackendSupportsMultimodal } from "../embedding-backend.js";
 import { asString, BackendUnavailableError } from "../job-utils.js";
 import { enqueueMemoryJob, type MemoryJob } from "../jobs-store.js";
@@ -22,11 +22,6 @@ const log = getLogger("memory-jobs-worker");
 
 export function rebuildIndexJob(): void {
   const db = getDb();
-  rawExec(/*sql*/ `DELETE FROM memory_segment_fts`);
-  rawExec(/*sql*/ `
-    INSERT INTO memory_segment_fts(segment_id, text)
-    SELECT id, text FROM memory_segments
-  `);
   db.delete(memoryEmbeddings).run();
 
   const items = db
