+++ assistant/src/memory/jobs-store.ts
@@ -25,7 +25,6 @@ export type MemoryJobType =
   | "build_conversation_summary"
   | "backfill"
   | "rebuild_index"
-  | "reconcile_fts"
   | "delete_qdrant_vectors"
   | "media_processing"
   | "embed_media"
@@ -317,23 +316,6 @@ export function enqueuePruneOldConversationsJob(
       ? { retentionDays }
       : {};
   return enqueueMemoryJob("prune_old_conversations", payload);
-}
-
-export function enqueueReconcileFtsJob(): string {
-  const db = getDb();
-  const existing = db
-    .select()
-    .from(memoryJobs)
-    .where(
-      and(
-        eq(memoryJobs.type, "reconcile_fts"),
-        inArray(memoryJobs.status, ["pending", "running"]),
-      ),
-    )
-    .orderBy(asc(memoryJobs.createdAt))
-    .get();
-  if (existing) return existing.id;
-  return enqueueMemoryJob("reconcile_fts", {});
 }
 
 export function claimMemoryJobs(limit: number): MemoryJob[] {
