+++ assistant/src/memory/retriever.test.ts
@@ -282,7 +282,6 @@ describe("Memory Retriever Degradation", () => {
     db.run("DELETE FROM memory_item_sources");
     db.run("DELETE FROM memory_items");
     db.run("DELETE FROM memory_segments");
-    db.run("DELETE FROM memory_segment_fts");
     db.run("DELETE FROM messages");
     db.run("DELETE FROM conversations");
     _resetQdrantBreaker();
