+++ assistant/src/tools/memory/handlers.test.ts
@@ -100,7 +100,6 @@ function clearTables() {
   db.run("DELETE FROM memory_item_sources");
   db.run("DELETE FROM memory_items");
   db.run("DELETE FROM memory_segments");
-  db.run("DELETE FROM memory_segment_fts");
   db.run("DELETE FROM messages");
   db.run("DELETE FROM conversations");
 }
